Create Encryption
This makes use of Fibonacci series logic to perform cryptography.
The program firstly consists of an encrypter that encrypts the saved text file. This encrypter uses a principle of Fibonacci series to implements its encryption phase. Each letter/character/number in the file is shifted according to its position on the file in reference to the Fibonacci series principle. This makes the message on the file unreadable to the perpetrator.